+++ src/test/java/net/jqwik/micronaut/ConstructorInjectionTest.java
@@ -1,11 +1,11 @@
 package net.jqwik.micronaut;
 
 import io.micronaut.context.annotation.Value;
+import org.assertj.core.api.SoftAssertions;
+
 import net.jqwik.api.Property;
 import net.jqwik.api.lifecycle.BeforeProperty;
-import net.jqwik.micronaut.annotation.JqwikMicronautTest;
 import net.jqwik.micronaut.beans.AppBean;
-import org.assertj.core.api.SoftAssertions;
 
 import static org.assertj.core.api.Assertions.assertThat;
