+++ src/test/java/net/jqwik/micronaut/DisableEmbeddedApplicationTest.java
@@ -2,8 +2,8 @@ package net.jqwik.micronaut;
 
 import io.micronaut.runtime.EmbeddedApplication;
 import jakarta.inject.Inject;
+
 import net.jqwik.api.Property;
-import net.jqwik.micronaut.annotation.JqwikMicronautTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
