+++ src/test/java/net/jqwik/micronaut/MathInnerService2Test.java
@@ -2,8 +2,8 @@ package net.jqwik.micronaut;
 
 import io.micronaut.test.annotation.MockBean;
 import jakarta.inject.Inject;
+
 import net.jqwik.api.Property;
-import net.jqwik.micronaut.annotation.JqwikMicronautTest;
 import net.jqwik.micronaut.beans.math.MathService;
 import net.jqwik.micronaut.beans.math.MathServiceImpl;
