+++ src/test/java/net/jqwik/micronaut/MathServiceTest.java
@@ -1,12 +1,12 @@
 package net.jqwik.micronaut;
 
 import jakarta.inject.Inject;
+
 import net.jqwik.api.Arbitraries;
 import net.jqwik.api.Arbitrary;
 import net.jqwik.api.ForAll;
 import net.jqwik.api.Property;
 import net.jqwik.api.Provide;
-import net.jqwik.micronaut.annotation.JqwikMicronautTest;
 import net.jqwik.micronaut.beans.math.MathService;
 
 import static org.assertj.core.api.Assertions.assertThat;
