+++ src/test/java/net/jqwik/micronaut/MathServiceTestSimilarNameTest.java
@@ -4,8 +4,8 @@ import io.micronaut.context.annotation.Requires;
 import io.micronaut.core.util.StringUtils;
 import io.micronaut.test.annotation.MockBean;
 import jakarta.inject.Inject;
+
 import net.jqwik.api.Property;
-import net.jqwik.micronaut.annotation.JqwikMicronautTest;
 import net.jqwik.micronaut.beans.math.MathService;
 import net.jqwik.micronaut.beans.math.MathServiceImpl;
