+++ src/test/java/net/jqwik/micronaut/NonTransactionalTest.java
@@ -3,9 +3,9 @@ package net.jqwik.micronaut;
 import io.micronaut.context.ApplicationContext;
 import io.micronaut.transaction.test.DefaultTestTransactionExecutionListener;
 import jakarta.inject.Inject;
+
 import net.jqwik.api.Property;
 import net.jqwik.micronaut.annotation.DbProperties;
-import net.jqwik.micronaut.annotation.JqwikMicronautTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
