+++ src/test/java/net/jqwik/micronaut/annotation/CustomBuilderMicronautTest.java
@@ -1,12 +1,13 @@
 package net.jqwik.micronaut.annotation;
 
-import net.jqwik.micronaut.beans.CustomContextBuilder;
-
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+
+import net.jqwik.micronaut.JqwikMicronautTest;
+import net.jqwik.micronaut.beans.CustomContextBuilder;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.METHOD, ElementType.ANNOTATION_TYPE, ElementType.TYPE})
